Send CreatePublisher rest message in HTTP channel on start
This is in response to a bug report that CreatePublisher was not being
called when the publisher is started. It was only being called when
we detected the publisher should be restarted.

As part of this work a basic unittest was added for the HTTP
channel based on the server code for the new rest agent.

 * Customer bug, CreatePublisher rest call was not always sent
 * Added basic unittest for http channel