Set PitLibraryPath in Peach Validator to current working folder
Currently the Peach Validator will change the current working folder
to the location of the pit file. This is in response to a bug in which
relative data filenames would not be found otherwise. However, once we
move there is no way to load a pit file as we will always have an
invalid PitLibraryPath