Support removing elements without cleanup
A common want is to move an element from one Block to antoher.
The natural assumption is to perform a .Remove and .Add.  However,
Remove -> RemoteAt -> OnRemoveItem will end up removing any bindings
(relations) from the element making it unusable.  This patch
adds ability to tell the api not to cleanup the item leaving
the Relations intact.  Also adds documentation and a unit test.

 - Update Remove, RemoteAt, and OnRemoveItem functions in
   DataElementContainer
 - Add DataELementCOntainerTests.cs with test for change
 - Update Frag element to utalize new API