Fix for PF-133
The ProcessMonitor, GdbDebugger, and CrashWrangler all got refactored to use a common platform-dependent Process class. When we spawn child processes, they might in turn spawn their own. If we only send a signal to the direct child of peach, we could end up waiting forever for the child to return after sending a SIGKILL if that child has children.

Fortunately, unix has the notion of a process group which allows for dispatching signals to a group of processes. When a shell composes a pipeline of commands, it uses a process group to manage the overall job.

So this new Process class arranges for the child to wind up as the leader of a new process group. Then when peach wants to kill the child, it uses killpg() to do so. Due to the way mono is implemented, we use a hack to call setpgid() at the appropriate time. The hack involves the use of a trampline process. It is very simple; it calls setpgid() and immediately calls execv().