Improve Process logic for usage by GDB.
* Add Attach(), Shutdown(), and Close() so that GdbDebugger can track the inferior.
* Add InferiorWaitForExit() for processes that were not directly created by the current process.
* Modify Stop() behavior so it now does:
  * Send SIGTERM
  * Wait for stdout/stderr to complete
  * If there was a timeout, send SIGKILL
  * Call WaitForExit to reap zombies (not exactly true under mono)
  * Close()

A further improvement would be to make the stdout/stderr wait be based on inactivity rather then a hard limit.

Additionally, the GdbDebugger now uses Shutdown() on the inferior (which sends a SIGTERM) and then waits for GDB to become idle. Hopefully this will allow GDB to finish running exploitable before we forcibly kill it.