Ask the TaskFactory to use a new thread for long running tasks.
By default, the task scheduler appears to interact badly with long running tasks.

The specific example was with the FTP_Client pit. It was trying to write to stdout and stderr, but apparently stderr was not getting consumed, causing the ftp script to deadlock. This in turn caused the TcpPublisher to timeout since we never got any input from the client.