Use a mono-based trampoline.
Instead of relying on native code or python, we can make use of mono's --debugger-agent.
It has an option to call setpgid(0, 0). We now have a simple trampoline written in C# and does not require any extra dependencies.

Also, define the peach-pro.zip and internal.zip in one place to minimize errors.