Added connect lifetime paramter to Tcp and TcpListener publishers
A need exists to perform multiple iterations of fuzzing over the
same TCP connection.  This is common in some SCADA protocols. We
have also had this requested by a customer.

A new parameter Lifetime was added with values of iteration or
session with iteration the default. When session is selected,
the connection will not be closed in OnClose.

 - Remvoed two asserts in BufferedStreamPublisher
 - Add CloseConnection call on send exception
 - Add new Lifetime parameter
 - Modify OnStart, OnStop, OnOpen and OnClose as needed