Removed parse test from pittester lint checks.
We already parse the pit several times. In the case of the lint check
parse the define variables are not setup the same as when testing the
pit which causes problems when using the TextFile and HexFile defines.

This check can be added back in post 3.8 ship if we would like, but
it will need some updates.