Completes PF-411
Pit authors may now add <FieldMask> as a child element of a <Data>. This can be used to constrain choices so that when generation fields tree metadata, only choices specified by <FieldMask> will be displayed. This should help deal with large pit data models seen in the wild, like NFSv4. This pit uses the same Data Model many times in eact output action. It's usually the case that there are a few choices that get made by <Field> selections during the output action.

The intent of <FieldMask> is to not make any change to the engine itself, only the generation of metadata should be effected.