Add `verifyAgainst` attr on PitTest output actions
Either verify output sent to the TestPublisher against explicit CDATA
in the test or the output action's DataFile (when available). The
latter scenario is useful for testing File-based pits where there are
no side-effects during a control iteration that would cause the
published output to differ from the source data set file.

Defaults to CDATA, so we'll need to explicitly set
`verifyAgainst='dataFile'` in the File pits.