Revert "Update to latest swashbuckle"
This reverts commit 27c165bfa3c27cc1d3e1fa752ac044ae94c44fa2.

The newer version of swashbuckle is build against .NET 4.0 and
references System.Net.Http 2.0.0.0.

mono is not letting us remap System.Net.Http 2.0.0.0 to 4.0.0.0.
With the appropriate entry in the app.config files, the mono
assembly loader fails to load Swashbuckle.Core.dll because
the dependent assembly can't be found.