PF-10: add back SnmpPowerMonitor
We'll support both SnmpPowerMonitor and ApcPowerMonitor. Under the hood
they share a lot of code. I made an abstract StartStopRestartMonitor
class for Monitors that support some notion of starting, stopping, and
restarting. It may also be useful for controlling the state of daemons.

Initially I tried making ApcPowerMonitor inherit directly from
SnmpPowerMonitor, but the Parameters were a mess. For instance, to get
the parameter parsing to work I'd have to expose the OIDs parameter on
the APC monitor, but the goal of that monitor is to hide OIDs.
Abstracting the underlying SnmpPowerDistributionUnit (used by
both monitors) as an IStartStopRestart for use by the new
StartStopRestartMonitor works much better.