Allow user to control fault/failure status codes in HTTP/REST
A common need during HTTP/REST fuzzing is to fault based on a
status code (such as 500) or to allow a status code that would
normally result in a failure (soft exception).

To this end the HttpPublisher and RestPublisher have been modified
to add two new parameters (FailureStatusCodes and FaultOnStatusCodes)
that allow the user to control both behaviours.  An initial list
of failure codes have been set to try and mimick the existing
behaviour.

To facilitate faulting from the Publisher a new property Context has
been added to the Publisher base class with a comment that the
property will be null when the publisher is remoted.  Changes to
the Engine were also made to set the Context for local publishers.

There is definetly a chance the Context property will be abused.
Not sure if there is a better option.