Bug: size relation unexpectedly dropped
The bug was discovered in the DICOM_Net_Provider pit, and manifests in
PitTester but not PeachValidator.

The test is about as minimal as I could make it. In the test,
'Response.Value.Items_0.A.Value' fails to crack because Peach claims
it's unsized. However, I expect the size relation on the sibling
'Length' (from the TLV ref) to make the 'Value' blob sized.

The debugger suggests the relation doesn't exist when the 'Value' is
getting cracked. It's possible a previous clone operation failed to
rewire the relations correctly.