PF-450: Store pits in .NET assembly
Pit state and data models may now be loaded from Peach.Pro.Pits.dll if
such a file is found in the root of the ##PitLibraryPath##. For now, the
keys for decrypting the pit contents are stored as a resource together
with Peach.Pro.dll. This will change as soon as Flexera licensing is
integrated.

* Replace License singleton with normal instance
* Move licensing code to License namespace