Redo control recording iteration fault handling PF-500
If a fault happens on a control recording iteration that was not the
first, it could be due to a previous fuzzing run.  When a fault is
detected on this type of iteration, we will now try and reproduce it.

If the fault reproduces on the first replay of the control recording
iteration, peach will stop fuzzing.

If the fault reprocudes on a non-first replay of the control recording
iteration, we record the search space, and continue fuzzing.  We will
rerun the control recording iteration again to ensure we get a clean
record before resuming fuzzing.

All soft exceptions on control record iterations are now treated as
faults so normal fault handling will kick in.

If we get a SoftException on control iteration fault, and the monitors
also detect a fault, the monitor fault will be prioritized higher for
bucketing. PF-501