POC: set context ID on every DICOM C-STORE outfrag
Requires a short-term hack in Frag.cs to bypass caching. It already
supports regenerating the default value when the payload is invalidated,
but currently DICOM has a static payload.

In DICOM the frag's template must be dynamically updated with the
negotiated context ID. Otherwise, it will likely send the wrong ID and
the target will close the session.