Initial checking of Peach Web doc tree
THis is a *very* early checkin with some basic structure
and existing documentation moved over.  There is allot of
content from Peach Fuzzer that has not been updated and no
doubt some build errors.

This does provide enough structure to start writing docs
for checks and asserts which is needed to close out their
JIRA tasks.